requirements.txt

streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
